backend/src/main/java/com/xaxis/disco/api/adapters/outbound/repository/handlers/UUIDTypeHandler.java
package com.xaxis.disco.api.adapters.outbound.repository.handlers;

import org.apache.ibatis.type.BaseTypeHandler;
import org.apache.ibatis.type.JdbcType;

import java.sql.CallableStatement;
import java.sql.PreparedStatement;
import java.sql.ResultSet;
import java.sql.SQLException;
import java.util.Objects;
import java.util.UUID;

import static org.apache.logging.log4j.util.Strings.isNotBlank;


public class UUIDTypeHandler extends BaseTypeHandler<UUID> {

    @Override
    public void setNonNullParameter(PreparedStatement ps, int i, UUID parameter, JdbcType jdbcType) throws SQLException {
        if (Objects.nonNull(parameter)) ps.setObject(i, parameter);
    }


    @Override
    public UUID getNullableResult(ResultSet rs, String columnName) throws SQLException {
        return convert(rs.getString(columnName));
    }

    @Override
    public UUID getNullableResult(ResultSet rs, int columnIndex) throws SQLException {
        return convert(rs.getString(columnIndex));
    }


    @Override
    public UUID getNullableResult(CallableStatement cs, int columnIndex) throws SQLException {
        return convert(cs.getString(columnIndex));
    }

    private UUID convert(String value) {
        return isNotBlank(value) ? UUID.fromString(value) : null;
    }


}
